+++ composeApp/src/commonMain/kotlin/emmanuelmuturia/sonux/repository/SonuxRepository.kt
@@ -1,0 +1,12 @@
+package emmanuelmuturia.sonux.repository
+
+import kotlinx.coroutines.CoroutineDispatcher
+
+expect class SonuxRepository {
+    suspend fun convertAndDownloadAudio(
+        inputUri: String,
+        frequency: Float = 0.49f,
+        coroutineDispatcher: CoroutineDispatcher
+    )
+}
+
