+++ composeApp/src/commonMain/kotlin/emmanuelmuturia/sonux/screens/ResultsScreen.kt
@@ -33,22 +33,17 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.rounded.ArrowBack
-import androidx.compose.material.icons.rounded.Add
-import androidx.compose.material.icons.rounded.PlayArrow
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
@@ -102,7 +97,8 @@ data class ResultsScreen(val audioFileUri: String) : Screen {
         ) { paddingValues ->
             ResultsScreenContent(
                 modifier = Modifier.padding(paddingValues = paddingValues),
-                sonuxViewModel = sonuxViewModel
+                sonuxViewModel = sonuxViewModel,
+                audioFileUri = audioFileUri
             )
         }
     }
@@ -111,9 +107,9 @@ data class ResultsScreen(val audioFileUri: String) : Screen {
 @Composable
 internal fun ResultsScreenContent(
     modifier: Modifier,
-    sonuxViewModel: SonuxViewModel
+    sonuxViewModel: SonuxViewModel,
+    audioFileUri: String
 ) {
-    val isPlaying by sonuxViewModel.isPlaying.collectAsState()
     LazyColumn(
         modifier = modifier.fillMaxSize().padding(all = 7.dp),
         verticalArrangement = Arrangement.Center,
@@ -134,9 +130,10 @@ internal fun ResultsScreenContent(
         item(
             key = "DownloadButton"
         ) {
-            PlayPauseButton(isPlaying = isPlaying) {
-                // sonuxViewModel.onPlayPauseButtonClicked(audioFileUri = audioFileUri)
-            }
+            DownloadButton(
+                sonuxViewModel = sonuxViewModel,
+                audioFileUri = audioFileUri
+            )
         }
 
         item(
@@ -173,36 +170,15 @@ internal fun ResultsConfirmationText() {
     Spacer(modifier = Modifier.height(height = 7.dp))
 }
 
-/*@Composable
-internal fun DownloadButton() {
+@Composable
+internal fun DownloadButton(sonuxViewModel: SonuxViewModel, audioFileUri: String) {
     Button(onClick = {
-        // Download the converted the audio file...
+        sonuxViewModel.convertAndDownloadAudio(audioFileUri = audioFileUri)
     }, colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.primary)) {
-        Text(text = "Download", style = MaterialTheme.typography.labelLarge)
+        Text(text = "Convert & Download", style = MaterialTheme.typography.labelLarge)
     }
 
     Spacer(modifier = Modifier.height(height = 7.dp))
-}*/
-
-@Composable
-internal fun PlayPauseButton(
-    isPlaying: Boolean, // Pass the playback state
-    onPlayPauseClicked: () -> Unit // Callback when button is clicked
-) {
-    IconButton(
-        onClick = { onPlayPauseClicked() }, // Trigger the play/pause function
-        colors = IconButtonDefaults.iconButtonColors(
-            containerColor = MaterialTheme.colorScheme.primary
-        )
-    ) {
-        Icon(
-            imageVector = if (isPlaying) Icons.Rounded.Add else Icons.Rounded.PlayArrow,
-            contentDescription = if (isPlaying) "Pause Button" else "Play Button",
-            tint = MaterialTheme.colorScheme.onPrimary
-        )
-    }
-
-    Spacer(modifier = Modifier.height(7.dp))
 }
 
 @Composable
