+++ composeApp/src/desktopMain/kotlin/emmanuelmuturia/sonux/repository/SonuxRepository.desktop.kt
@@ -1,0 +1,12 @@
+package emmanuelmuturia.sonux.repository
+
+import kotlinx.coroutines.CoroutineDispatcher
+
+actual class SonuxRepository {
+    actual suspend fun convertAndDownloadAudio(
+        inputUri: String,
+        frequency: Float,
+        coroutineDispatcher: CoroutineDispatcher
+    ) {
+    }
+}
